feat(editor): 部屋別 loot をスポーンテーブルと同じ割合バーで見せる
WeightBar をスポーンテーブルから共有コンポーネントへ切り出し、部屋別 loot ページでも
使う。期待個数の割合を横積みバーで色分けし、上位12種＋その他に畳んで読みやすくする。
表だけでは分かりにくい構成比を一目で掴めるようにする。

- editor-ui/src/components/WeightBar.tsx: 新設。TableViewer から抽出した共有バー
- editor-ui/src/pages/TableViewerPage.tsx: ローカル WeightBar を共有版に置換
- editor-ui/src/pages/RoomLootPage.tsx: 期待個数の割合バーを表の上に追加
